+++ OnlineCourse.Core/WorkFlows/LoginUser/Validation.cs
@@ -36,10 +36,10 @@ namespace OnlineCourse.Core.WorkFlows.LoginUser
                     if (user.ValidMobile != (byte) ValidationState.Valid)
                         return (byte) LoginUserMessage.AccountNotVerified;
 
-                    if (user.State != (byte)UserState.Approved && (user.AccessLevel==UserAccessLevel.Administrator  || user.AccessLevel ==UserAccessLevel.Teacher))
+                    if (user.State != UserState.Approved && (user.AccessLevel==UserAccessLevel.Administrator  || user.AccessLevel ==UserAccessLevel.Teacher))
                         return (byte)LoginUserMessage.AccountDisabled;
 
-                    if (user.State != (byte)UserState.Verified && user.State != (byte)UserState.Approved)
+                    if (user.State != UserState.Verified && user.State != UserState.Approved)
                         return (byte)LoginUserMessage.AccountNotVerified;
 
                     if (!PublicValidator.IpCheck(login.Ip))
